requirements.txt
aerich==0.9.2
aiogram==3.22.0
aiogram-tests==1.0.3
aiohttp==3.12.15
asyncpg==0.30.0
pydantic==2.11.10
pydantic-settings==2.11.0
python-dotenv==1.1.1
tortoise-orm==0.25.1
aiogram_i18n==1.4
fluent.runtime==0.4.0

